requirements.txt
# F1 Motion Simulator Dependencies
# Install with: pip install -r requirements.txt

# Serial communication (Modbus for SMC)
pymodbus>=3.0.0
pyserial>=3.5

# Configuration
PyYAML>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development (optional)
black>=23.0.0      # Code formatting
flake8>=6.0.0      # Linting
mypy>=1.0.0        # Type checking

# Logging (optional, for colored output)
colorlog>=6.0.0
